+++ include/GGE/Core/maps/Tileset.hpp
@@ -16,17 +16,45 @@ public:
 	Tileset();
 	~Tileset();
 
+	/**
+	 * Registers the application described by theApp.
+	 * 
+	 * @param [in,out]	theApp	If non-null, the application.
+	 */
 	void RegisterApp(App* theApp);
+
+	/**
+	 * Inicializa el tileset.
+	 *
+	 * @param [in,out]	TheTmx	Puntero al archivo tmx.
+	 */
 	void Init(TmxMap* TheTmx);
-	void Draw(int theIndx, sf::Vector2f thePos);
+
+	/**
+	 * Dibuja un tile en pantalla
+	 *
+	 * @param	theIndx	el indice del tile que se quiere dibujar.
+	 * @param	thePos 	Vector con la posición en la que se dibujará.
+	 */
+	void Draw(int theIndx, sf::Vector2i thePos);
 
 private:
 	/// Puntero a APP
 	App* mApp;
+	/// Vector que contiene las imágenes
 	std::vector<sf::Sprite*> mImages;
+	/// Vector que contiene el número del primer tile del tileset
 	std::vector<int> mFirstgid;
+	/// Vector que contiene los rets de los tilesets
 	std::vector<sf::IntRect> mRects;
 
+	/**
+	 * Devuelve el tileset al que pertenece un tile en concreto.
+	 *
+	 * @param	theIndx	the indx.
+	 *
+	 * @return	The number tileset.
+	 */
 	int GetNumTileset(int theIndx) const;
